rpc: Auto-derive the ValueHint trait
