+++ chainstate/Cargo.toml
@@ -15,6 +15,7 @@ crypto = {path = '../crypto'}
 logging = {path = '../logging'}
 pos-accounting = {path = '../pos-accounting'}
 rpc = {path = '../rpc'}
+rpc-description = {path = '../rpc/description'}
 serialization = {path = "../serialization"}
 subsystem = {path = '../subsystem'}
 tx-verifier = {path = './tx-verifier'}
