+++ chainstate/src/detail/info.rs
@@ -19,23 +19,12 @@ use common::{
     chain::{block::timestamp::BlockTimestamp, GenBlock},
     primitives::{BlockHeight, Id},
 };
-use rpc::description::ValueHint as VH;
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, rpc::description::HasValueHint)]
 pub struct ChainInfo {
     pub best_block_height: BlockHeight,
     pub best_block_id: Id<GenBlock>,
     pub best_block_timestamp: BlockTimestamp,
     pub median_time: BlockTimestamp,
     pub is_initial_block_download: bool,
-}
-
-impl rpc::description::HasValueHint for ChainInfo {
-    const HINT: VH = VH::Object(&[
-        ("best_block_height", &BlockHeight::HINT),
-        ("best_block_id", &<Id<GenBlock>>::HINT),
-        ("best_block_timestamp", &BlockTimestamp::HINT),
-        ("median_time", &BlockTimestamp::HINT),
-        ("is_initial_block_download", &bool::HINT),
-    ]);
 }
