+++ chainstate/src/rpc/types/event.rs
@@ -17,11 +17,10 @@ use common::{
     chain::Block,
     primitives::{BlockHeight, Id},
 };
-use rpc::description::ValueHint as VH;
 
 use crate::ChainstateEvent;
 
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, rpc::description::HasValueHint)]
 pub enum RpcEvent {
     NewTip { id: Id<Block>, height: BlockHeight },
 }
@@ -32,11 +31,4 @@ impl RpcEvent {
             ChainstateEvent::NewTip(id, height) => Self::NewTip { id, height },
         }
     }
-}
-
-impl rpc::description::HasValueHint for RpcEvent {
-    const HINT: VH = VH::Object(&[(
-        "NewTip",
-        &VH::Object(&[("id", &<Id<Block>>::HINT), ("height", &BlockHeight::HINT)]),
-    )]);
 }
