+++ mempool/src/pool/feerate.rs
@@ -17,7 +17,7 @@ use serde::{Deserialize, Serialize};
 use std::num::NonZeroUsize;
 
 use common::primitives::amount::Amount;
-use rpc::description::ValueHint as VH;
+use rpc::description::HasValueHint;
 
 use crate::error::MempoolPolicyError;
 
@@ -28,7 +28,9 @@ pub const INCREMENTAL_RELAY_FEE_RATE: FeeRate =
 pub const INCREMENTAL_RELAY_THRESHOLD: FeeRate =
     FeeRate::from_amount_per_kb(Amount::from_atoms(500));
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
+#[derive(
+    Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize, HasValueHint,
+)]
 pub struct FeeRate {
     amount_per_kb: Amount,
 }
@@ -72,10 +74,6 @@ impl std::ops::Add for FeeRate {
     fn add(self, other: Self) -> Self::Output {
         (self.amount_per_kb + other.amount_per_kb).map(FeeRate::from_amount_per_kb)
     }
-}
-
-impl rpc::description::HasValueHint for FeeRate {
-    const HINT: VH = VH::Object(&[("amount_per_kb", &VH::NUMBER)]);
 }
 
 #[cfg(test)]
