+++ wallet/types/src/with_locked.rs
@@ -13,17 +13,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use rpc_description::ValueHint as VH;
+use rpc_description::HasValueHint;
 
 /// Enum used to specify whether to include locked balance/utxos for wallet commands
-#[derive(Debug, Clone, Copy, serde::Serialize, serde::Deserialize)]
+#[derive(Debug, Clone, Copy, serde::Serialize, serde::Deserialize, HasValueHint)]
 pub enum WithLocked {
     Any,
     Unlocked,
     Locked,
-}
-
-impl rpc_description::HasValueHint for WithLocked {
-    const HINT: VH =
-        VH::Choice(&[&VH::StrLit("Any"), &VH::StrLit("Unlocked"), &VH::StrLit("Locked")]);
 }
