+++ wallet/wallet-rpc-lib/Cargo.toml
@@ -15,6 +15,7 @@ mempool-types = { path = "../../mempool/types" }
 mempool = { path = "../../mempool" }
 node-comm = { path = "../wallet-node-client" }
 rpc = { path = "../../rpc" }
+rpc-description = { path = "../../rpc/description" }
 serialization = { path = "../../serialization" }
 utils = { path = "../../utils" }
 utils-networking = { path = "../../utils/networking" }
